requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
arabictransliterator @ git+https://github.com/mtg/ArabicTransliterator.git@df0296c5
    # via -r requirements.in
asgiref==3.7.2
    # via django
beautifulsoup4==4.12.2
    # via -r requirements.in
billiard==4.2.0
    # via celery
cachetools==5.5.0
    # via google-auth
celery==5.4.0
    # via -r requirements.in
certifi==2019.3.9
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.1
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
contourpy==1.1.1
    # via matplotlib
coverage[toml]==5.3.1
    # via eyed3
cycler==0.10.0
    # via matplotlib
decorator==4.4.0
    # via networkx
deprecation==2.1.0
    # via eyed3
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-manifest-loader
    #   django-sendfile2
    #   djangorestframework
django-manifest-loader==1.0.0
    # via -r requirements.in
django-sendfile2==0.7.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
eyed3==0.9.7
    # via
    #   -r requirements.in
    #   pycompmusic
filetype==1.2.0
    # via
    #   -r requirements.in
    #   eyed3
fonttools==4.43.1
    # via matplotlib
google-api-core==2.21.0
    # via google-api-python-client
google-api-python-client==2.94.0
    # via -r requirements.in
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.0.0
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via google-api-core
httplib2==0.18.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.8
    # via requests
intonation==0.1
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
kombu==5.3.7
    # via celery
levenshtein==0.25.1
    # via python-levenshtein
matplotlib==3.9.0
    # via
    #   -r requirements.in
    #   intonation
    #   pypeaks
    #   seyiranalyzer
musicbrainzngs==0.7.1
    # via
    #   -r requirements.in
    #   pycompmusic
mwparserfromhell==0.5.4
    # via -r requirements.in
networkx==2.3
    # via -r requirements.in
numpy==1.26.1
    # via
    #   -r requirements.in
    #   contourpy
    #   intonation
    #   matplotlib
    #   pypeaks
    #   scipy
    #   seyiranalyzer
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via
    #   deprecation
    #   matplotlib
pillow==10.3.0
    # via
    #   -r requirements.in
    #   matplotlib
prompt-toolkit==2.0.9
    # via click-repl
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
pycompmusic @ git+https://github.com/mtg/pycompmusic.git@72b689e
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pyparsing==2.4.0
    # via matplotlib
pypeaks==0.2.7
    # via intonation
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   celery
    #   matplotlib
python-levenshtein==0.25.1
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
pytz==2024.1
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
rapidfuzz==3.9.3
    # via levenshtein
redis==5.0.6
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   pycompmusic
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scipy==1.13.1
    # via
    #   -r requirements.in
    #   intonation
    #   seyiranalyzer
sentry-sdk==2.6.0
    # via -r requirements.in
seyiranalyzer @ git+https://github.com/sertansenturk/seyiranalyzer.git@04392597
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   cycler
    #   google-auth-httplib2
    #   prompt-toolkit
    #   pycompmusic
    #   python-dateutil
soupsieve==2.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
tomato @ git+https://github.com/sertansenturk/tomato.git@v0.13.0
    # via -r requirements.in
toml==0.10.2
    # via coverage
typing-extensions==4.8.0
    # via dj-database-url
tzdata==2023.3
    # via celery
unidecode==1.3.8
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.0.7
    # via
    #   requests
    #   sentry-sdk
uwsgi==2.0.26
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.1.7
    # via prompt-toolkit
whitenoise==6.7.0
    # via -r requirements.in
